+++ lib/app.rb
@@ -1,6 +1,7 @@
 module App
   class Error < StandardError; end
   class ModeNotFound < Error; end
+  class SystemNotFound < Error; end
 
   #
 
@@ -17,6 +18,7 @@ module App
     setup_env
     setup_bundle
     setup_liza
+    bundle_systems_app Lizarb::APP_DIR
 
     check_mode!
 
@@ -33,7 +35,7 @@ module App
 
   def setup_bundle
     require "bundler/setup"
-    Bundler.require :default
+    Bundler.require :default, *@systems.keys
   end
 
   def setup_liza
@@ -48,6 +50,36 @@ module App
 
     loader.enable_reloading
     loader.setup
+  end
+
+  def bundle_systems_app app_dir
+    @systems.keys.each do |k|
+      key = "#{k}_system"
+
+      require_system key
+      klass = Object.const_get key.camelize
+
+      @systems[k] = klass
+    end
+
+    @loaders << loader = Zeitwerk::Loader.new
+
+    @systems.each do |k, klass|
+      # ORDER MATTERS: IGNORE, COLLAPSE, PUSH
+      loader.collapse "#{fname_for klass}/**/*"
+      loader.push_dir "#{fname_for klass}", namespace: klass
+    end
+
+    # ORDER MATTERS: IGNORE, COLLAPSE, PUSH
+    loader.collapse "#{app_dir}/app/**/*"
+    loader.push_dir "#{app_dir}/app" if Dir.exist? "#{app_dir}/app"
+
+    loader.enable_reloading
+    loader.setup
+
+    @systems.each do |k, klass|
+      connect_system k, klass
+    end
   end
 
   # loaders
@@ -82,6 +114,26 @@ module App
   def check_mode!
     return if @modes.include? @mode
     raise ModeNotFound, "LIZA_MODE `#{@mode}` not found in #{@modes}", []
+  end
+
+  # systems
+
+  @systems = {}
+
+  def system key
+    raise "locked" if @locked
+    @systems[key] = nil
+  end
+
+  def systems
+    @systems
+  end
+
+  def self.require_system key
+    require key
+  rescue LoadError => e
+    def e.backtrace; []; end
+    raise SystemNotFound, "FILE #{key}.rb not found on $LOAD_PATH", []
   end
 
   # dev
@@ -140,8 +192,13 @@ module App
 
   # parts
 
-  def connect_part part_klass, key
-    klass = Liza.const "#{key}_part"
+  def connect_part part_klass, key, system
+    klass = if system.nil?
+              Liza.const "#{key}_part"
+            else
+              Liza.const("#{system}_system")
+                  .const "#{key}_part"
+            end
 
     log "CONNECTING PART #{part_klass.to_s.rjust 25}.part :#{key}"
 
@@ -153,6 +210,34 @@ module App
       klass.const_set :Extension, Class.new(Liza::PartExtension)
       klass::Extension.class_exec &klass.extension
     end
+  end
+
+  # systems
+
+  def connect_system key, system_klass
+    t = Time.now
+
+    color_system_klass = system_klass.to_s.colorize system_klass.log_color
+    color_key = key.to_s.colorize system_klass.log_color
+
+    registrar_index = 0
+    system_klass.registrar.each do |string, target_block|
+      reg_type, _sep, reg_target = string.to_s.lpartition "_"
+
+      registrar_index += 1
+
+      target_klass = Liza.const reg_target
+
+      if reg_type == "insertion"
+        target_klass.class_exec &target_block
+      else
+        raise "TODO: decide and implement system extension"
+      end
+
+      log "CONNECTING SYSTEM PART          #{color_system_klass}.#{reg_type} #{target_klass}"
+
+    end
+    log "CONNECTING SYSTEM - #{t.diff}s for #{color_system_klass} to connect to #{registrar_index} system parts"
   end
 
   #