+++ lib/liza/meta/system_test.rb
@@ -1,0 +1,13 @@
+module Liza
+  class SystemTest < UnitTest
+    test :subject_class do
+      assert subject_class == Liza::System
+    end
+
+    test :settings do
+      assert subject_class.log_level == :normal
+      assert subject_class.log_color == :white
+    end
+
+  end
+end