+++ CHANGELOG.md
@@ -5,4 +5,4 @@
 - Liza::Test
 - Liza::Unit has class/instance methods settings/set/set/add/fetch
 - Liza::Unit has class/instance methods log/log_color/log_level/log?/log_level?
-
+- Liza::System, Liza::Box, Liza::Panel, Liza::Controller